requirements.txt
appdirs==1.4.4
Pillow==7.1.2
selenium==3.141.0
urllib3==1.25.9
virtualenv==20.0.30
pandas==1.0.3
matplotlib==3.3.3
pathlib==1.0.1
